+++ samples/Sample_02_ReadGanttProject.php
@@ -27,46 +27,16 @@ echo EOL;
 
 // Ressources
 echo date('H:i:s') . ' Get ressources'.EOL;
-$oResources = $objPHPProject->getAllResources();
-foreach ($oResources as $item){
-    echo 'Resource : '.$item->getTitle().EOL;
+foreach ($objPHPProject->getAllResources() as $oResource){
+    echo 'Resource : '.$oResource->getTitle().EOL;
 }
 echo EOL;
 
 // Tasks
 echo date('H:i:s') . ' Get tasks'.EOL;
-$oTasks = $objPHPProject->getAllTasks();
-foreach ($oTasks as $item){
-    echo 'Task : '.$item->getName().EOL;
-    echo ' >> Duration : '.$item->getDuration().EOL;
-    echo ' >> StartDate : '.date('Y-m-d', $item->getStartDate()).EOL;
-    echo ' >> Progress : '.$item->getProgress().EOL;
-    echo ' >> Resources : '.EOL;
-    $oTaskResources = $item->getResources();
-    if(!empty($oTaskResources)){
-        foreach ($oTaskResources as $itemRes){
-            echo ' >>>> Resource : '.$objPHPProject->getResource($itemRes)->getTitle().EOL;
-        }
-    }
-    
-    echo ' >> SubTasks : '.EOL;
-    if($item->getTaskCount() > 0){
-        foreach ($item->getTasks() as $itemSub){
-            echo ' >>>> Task : '.$itemSub->getName().EOL;
-            echo ' >>>>>> Duration : '.$itemSub->getDuration().EOL;
-            echo ' >>>>>> StartDate : '.date('Y-m-d', $itemSub->getStartDate()).EOL;
-            echo ' >>>>>> Progress : '.$itemSub->getProgress().EOL;
-            echo ' >>>>>> Resources : '.EOL;
-            $oTaskResources = $itemSub->getResources();
-            if(!empty($oTaskResources)){
-                foreach ($oTaskResources as $itemRes){
-                    echo ' >>>>>>>> Resource : '.$objPHPProject->getResource($itemRes)->getTitle().EOL;
-                }
-            }
-        }
-    }
+foreach ($objPHPProject->getAllTasks() as $oTask){
+	echoTask($objPHPProject, $oTask);
 }
-echo EOL;
 
 // Echo done
 echo date('H:i:s') . ' Done reading file.'.EOL;