+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.2.0 - 2014-08-13
+### Features
+- MSProjectExchange Reader - @Progi1984 GH-4
+- MSProjectExchange Writer - @Progi1984 GH-2
+
+### Miscellaneous
+- Refactored resources management - @Progi1984
+
+
 ## 0.1.0 - 2014-08-08
 
 ### Features
@@ -7,8 +16,6 @@
 - Support of namespaces - @Progi1984 GH-12
 - GanttProject Writer - @Progi1984 GH-1
 - GanttProject Reader - @Progi1984 GH-3
-
-### Bugfix
 
 ### Miscellaneous
 - QA : Documentation - @Progi1984 GH-8 GH-12